requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Mail
Flask-Migrate
PyMySQL
cryptography
python-dotenv
Pillow
WTForms
email-validator
gunicorn
